优化占用
